Big refactor based on PR feedback: don’t put function in redux store, keep data requesting closer to data usage
