+++ amp/app/containers/home/container.jsx
@@ -7,6 +7,8 @@ import Link from '../../components/link'
 
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import {getHomeCategories} from '../../../../web/app/containers/home/selectors'
+import {initHomePage} from '../../../../web/app/integration-manager/home/commands'
+import {CURRENT_URL} from '../../../../web/app/containers/app/constants'
 
 // import container styles
 import containerStyles from './container.scss'
@@ -29,9 +31,11 @@ Home.propTypes = {
     categories: PropTypes.array.isRequired
 }
 
+Home.resolves = [({dispatch, getState}) => {
+    return dispatch(initHomePage(getState().ui.app.get(CURRENT_URL)))
+}]
+
 Home.templateName = 'home'
-
-
 
 const mapStateToProps = createPropsSelector({
     categories: getHomeCategories
