+++ amp/app/data-integration/constants.js
@@ -1,2 +1,1 @@
 export const PAGE_TITLE = 'pageTitle'
-export const DATA_INIT_FUNCTION = 'dataInitFunction'
