+++ amp/package.json
@@ -10,7 +10,6 @@
   },
   "devDependencies": {
     "amphtml-validator": "^1.0.20",
-    "archiver": "1.3.0",
     "autoprefixer": "6.7.7",
     "babel-cli": "6.24.1",
     "babel-loader": "6.4.1",
